+++ src/components/DuplicantMenu/DuplicantMenu.tsx
@@ -1,10 +1,13 @@
 import * as React from "react";
+import { connect } from "react-redux";
 
 import IconButton from "@material-ui/core/IconButton";
 import Menu from "@material-ui/core/Menu";
 import MenuItem from "@material-ui/core/MenuItem";
+import Divider from "@material-ui/core/Divider";
 
 import MoreVertIcon from "@material-ui/icons/MoreVert";
+
 import AbstractCopyButton from "@/services/oni-save/components/AbstractCopyButton";
 import {
   AccessorizerBehavior,
@@ -12,12 +15,14 @@ import {
 } from "oni-save-parser";
 import AbstractPasteButton from "@/services/oni-save/components/AbstractPasteButton";
 
+import mapDispatchToProps, { DispatchProps } from "./dispatch-props";
+
 export interface DuplicantMenuProps {
   gameObjectId: number;
 }
 
-type Props = DuplicantMenuProps;
-const DuplicantMenu: React.SFC<Props> = ({ gameObjectId }) => {
+type Props = DuplicantMenuProps & DispatchProps;
+const DuplicantMenu: React.SFC<Props> = ({ gameObjectId, cloneDuplicant }) => {
   const [anchorEl, setAnchorEl] = React.useState<HTMLElement | null>(null);
   function onClose() {
     setAnchorEl(null);
@@ -72,9 +77,21 @@ const DuplicantMenu: React.SFC<Props> = ({ gameObjectId }) => {
             </MenuItem>
           )}
         </AbstractPasteButton>
+        <Divider />
+        <MenuItem
+          onClick={() => {
+            onClose();
+            cloneDuplicant(gameObjectId);
+          }}
+        >
+          Clone
+        </MenuItem>
       </Menu>
     </div>
   );
 };
 
-export default DuplicantMenu;
+export default connect(
+  null,
+  mapDispatchToProps
+)(DuplicantMenu);