+++ src/components/DuplicantMenu/dispatch-props.ts
@@ -1,0 +1,7 @@
+import { cloneDuplicant } from "@/services/oni-save/actions/clone-duplicant";
+
+const mapDispatchToProps = {
+  cloneDuplicant
+};
+export type DispatchProps = typeof mapDispatchToProps;
+export default mapDispatchToProps;