+++ src/services/oni-save/reducer/index.ts
@@ -1,5 +1,6 @@
 import reduceReducers from "reduce-reducers";
 
+import cloneDuplicantReducer from "./clone-duplicant";
 import copyBehaviorsReducer from "./copy-behaviors";
 import modifyBehaviorReducer from "./modify-behavior";
 import parseProgressReducer from "./parse-progress";
@@ -7,6 +8,7 @@ import pasteBehaviorsReducer from "./paste-behaviors";
 import receiveOniSaveReducer from "./receive-onisave";
 
 export default reduceReducers(
+  cloneDuplicantReducer,
   copyBehaviorsReducer,
   modifyBehaviorReducer,
   parseProgressReducer,